kernel-sidecar/requirements.txt
pycryptodomex
kubernetes 
jupyter_client
